requirements.txt
pillow
requests==2.26.0
Gooey